+++ tests/tests_pytorch/trainer/connectors/test_accelerator_connector.py
@@ -41,7 +41,6 @@ from lightning.pytorch.accelerators.mps import MPSAccelerator
 from lightning.pytorch.plugins import DoublePrecisionPlugin, LayerSync, PrecisionPlugin, TorchSyncBatchNorm
 from lightning.pytorch.plugins.io import TorchCheckpointIO
 from lightning.pytorch.strategies import (
-    DDPSpawnStrategy,
     DDPStrategy,
     DeepSpeedStrategy,
     FSDPStrategy,
@@ -51,7 +50,7 @@ from lightning.pytorch.strategies import (
     SingleTPUStrategy,
     XLAStrategy,
 )
-from lightning.pytorch.strategies.ddp_spawn import _DDP_FORK_ALIASES
+from lightning.pytorch.strategies.ddp import _DDP_FORK_ALIASES
 from lightning.pytorch.strategies.hpu_parallel import HPUParallelStrategy
 from lightning.pytorch.trainer.connectors.accelerator_connector import _set_torch_flags, AcceleratorConnector
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
@@ -96,12 +95,6 @@ def test_invalid_strategy_choice(invalid_strategy):
 def test_strategy_choice_ddp_on_cpu(tmpdir):
     """Test that selecting DDPStrategy on CPU works."""
     _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class=DDPStrategy)
-
-
-@RunIf(skip_windows=True)
-def test_strategy_choice_ddp_spawn_on_cpu(tmpdir):
-    """Test that selecting DDPSpawnStrategy on CPU works."""
-    _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class=DDPSpawnStrategy)
 
 
 def _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class):
@@ -254,7 +247,7 @@ def test_interactive_compatible_strategy_ddp_fork(monkeypatch):
     ["strategy", "strategy_class"],
     [
         ("ddp", DDPStrategy),
-        ("ddp_spawn", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
         pytest.param("deepspeed", DeepSpeedStrategy, marks=RunIf(deepspeed=True)),
     ],
 )
@@ -332,9 +325,9 @@ def test_set_devices_if_none_cpu():
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     (
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_true", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_true", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
         ("ddp_find_unused_parameters_true", DDPStrategy),
@@ -353,8 +346,8 @@ def test_invalid_ddp_strategy_with_mps(accelerator, strategy, strategy_class, mp
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     [
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
     ],
@@ -364,17 +357,16 @@ def test_strategy_choice_cpu_str(strategy, strategy_class):
     assert isinstance(trainer.strategy, strategy_class)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_strategy_choice_cpu_instance(strategy_class):
-    trainer = Trainer(strategy=strategy_class(), accelerator="cpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_strategy_choice_cpu_instance():
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="cpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
 
 
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     [
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
         pytest.param("deepspeed", DeepSpeedStrategy, marks=RunIf(deepspeed=True)),
@@ -385,16 +377,14 @@ def test_strategy_choice_gpu_str(strategy, strategy_class, cuda_count_2, mps_cou
     assert isinstance(trainer.strategy, strategy_class)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_strategy_choice_gpu_instance(strategy_class, cuda_count_2, mps_count_0):
-    trainer = Trainer(strategy=strategy_class(), accelerator="gpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_strategy_choice_gpu_instance(cuda_count_2, mps_count_0):
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="gpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_device_type_when_strategy_instance_gpu_passed(strategy_class, cuda_count_2, mps_count_0):
-    trainer = Trainer(strategy=strategy_class(), accelerator="gpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_device_type_when_strategy_instance_gpu_passed(cuda_count_2, mps_count_0):
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="gpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.accelerator, CUDAAccelerator)
 
 
@@ -408,7 +398,7 @@ def test_validate_precision_type(precision):
 def test_strategy_choice_ddp_spawn_cpu():
     trainer = Trainer(strategy="ddp_spawn", accelerator="cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "spawn"
 
@@ -420,7 +410,7 @@ def test_strategy_choice_ddp_fork_in_interactive():
     environments by default."""
     trainer = Trainer(devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "fork"
 
@@ -429,12 +419,12 @@ def test_strategy_choice_ddp_fork_in_interactive():
 def test_strategy_choice_ddp_fork_cpu():
     trainer = Trainer(strategy="ddp_fork", accelerator="cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "fork"
 
 
-@pytest.mark.parametrize("strategy,expected_cls", [("ddp", DDPStrategy), ("ddp_spawn", DDPSpawnStrategy)])
+@pytest.mark.parametrize("strategy,expected_cls", [("ddp", DDPStrategy), ("ddp_spawn", DDPStrategy)])
 def test_strategy_choice_ddp_cuda(strategy, expected_cls, mps_count_0, cuda_count_2):
     trainer = Trainer(fast_dev_run=True, strategy=strategy, accelerator="gpu", devices=1)
     assert isinstance(trainer.accelerator, CUDAAccelerator)